Create requirements
To download all required modules in order for scrapecmd to work,  use the requirements.txt file.
in the cmd line: "pip install -r path/to/requirements.txt" where path is the directory that contains the requirements file. -r is a shorter version of --requirement which means "download from requirements file".